Disable cursorline and hidden
Disable cursorline because it would be necessary to modify the
colorscheme to highlight the number column.

Also disable hidden to prevent losing work from hidden buffers.
The side-effect of this option is that it won't be possible to let
things run in a hidden terminal window, since it will kill the terminal.